+++ packages/calcite-components/src/components/accordion-item/accordion-item.tsx
@@ -30,6 +30,8 @@ declare global {
  * @slot - A slot for adding custom content, including nested `calcite-accordion-item`s.
  * @slot actions-end - A slot for adding `calcite-action`s or content to the end side of the component's header.
  * @slot actions-start - A slot for adding `calcite-action`s or content to the start side of the component's header.
+ * @slot content-end - A slot for adding non-actionable elements after the component's header text.
+ * @slot content-start - A slot for adding non-actionable elements before the component's header text.
  */
 export class AccordionItem extends LitElement {
   //#region Static Members
@@ -58,6 +60,10 @@ export class AccordionItem extends LitElement {
   @state() hasActionsEnd = false;
 
   @state() hasActionsStart = false;
+
+  @state() hasContentEnd = false;
+
+  @state() hasContentStart = false;
 
   //#endregion
 
@@ -247,6 +253,14 @@ export class AccordionItem extends LitElement {
     this.hasActionsEnd = slotChangeHasAssignedElement(event);
   }
 
+  private handleContentEndSlotChange(event: Event): void {
+    this.hasContentEnd = slotChangeHasAssignedElement(event);
+  }
+
+  private handleContentStartSlotChange(event: Event): void {
+    this.hasContentStart = slotChangeHasAssignedElement(event);
+  }
+
   /** handle clicks on item header */
   private itemHeaderClickHandler(): void {
     this.emitRequestedItem();
@@ -295,6 +309,22 @@ export class AccordionItem extends LitElement {
     return (
       <div class={CSS.actionsEnd} hidden={!this.hasActionsEnd}>
         <slot name={SLOTS.actionsEnd} onSlotChange={this.handleActionsEndSlotChange} />
+      </div>
+    );
+  }
+
+  private renderContentEnd(): JsxNode {
+    return (
+      <div class={CSS.slotContentEnd} hidden={!this.hasContentEnd}>
+        <slot name={SLOTS.contentEnd} onSlotChange={this.handleContentEndSlotChange} />
+      </div>
+    );
+  }
+
+  private renderContentStart(): JsxNode {
+    return (
+      <div class={CSS.slotContentStart} hidden={!this.hasContentStart}>
+        <slot name={SLOTS.contentStart} onSlotChange={this.handleContentStartSlotChange} />
       </div>
     );
   }
@@ -349,6 +379,7 @@ export class AccordionItem extends LitElement {
             tabIndex="0"
           >
             <div class={CSS.headerContainer}>
+              {this.renderContentStart()}
               {iconStartEl}
               <div class={CSS.headerText}>
                 <Heading class={CSS.heading} level={headingLevel}>
@@ -357,6 +388,7 @@ export class AccordionItem extends LitElement {
                 {description ? <span class={CSS.description}>{description}</span> : null}
               </div>
               {iconEndEl}
+              {this.renderContentEnd()}
             </div>
             <calcite-icon
               class={CSS.expandIcon}